requirements.txt
torch>=1.7.1
opencv-python
numpy
Pillow
customtkinter
python-vlc
ultralytics
